Trait casper_dao_contracts::bid_escrow::BidEscrowContractInterface

pub trait BidEscrowContractInterface {

    fn init(
        &mut self,
        variable_repository: Address,
        reputation_token: Address,
        kyc_token: Address,
        va_token: Address
    );
    fn post_job_offer(
        &mut self,
        expected_timeframe: BlockTime,
        budget: U512,
        purse: URef
    );
    fn submit_bid(
        &mut self,
        job_offer_id: JobOfferId,
        time: BlockTime,
        payment: U512,
        reputation_stake: U512,
        onboard: bool,
        purse: Option<URef>
    );
    fn cancel_bid(&mut self, bid_id: BidId);
    fn pick_bid(&mut self, job_offer_id: u32, bid_id: u32, purse: URef);
    fn submit_job_proof(&mut self, job_id: JobId, proof: DocumentHash);
    fn submit_job_proof_during_grace_period(
        &mut self,
        job_id: JobId,
        proof: DocumentHash,
        reputation_stake: U512,
        onboard: bool,
        purse: Option<URef>
    );
    fn vote(
        &mut self,
        voting_id: VotingId,
        voting_type: VotingType,
        choice: Choice,
        stake: U512
    );
    fn finish_voting(&mut self, voting_id: VotingId, voting_type: VotingType);
    fn get_job(&self, job_id: JobId) -> Option<Job>;
    fn get_job_offer(&self, job_offer_id: JobOfferId) -> Option<JobOffer>;
    fn get_bid(&self, bid_id: BidId) -> Option<Bid>;
    fn variable_repository_address(&self) -> Address;
    fn reputation_token_address(&self) -> Address;
    fn get_voting(&self, voting_id: VotingId) -> Option<VotingStateMachine>;
    fn get_ballot(
        &self,
        voting_id: VotingId,
        voting_type: VotingType,
        address: Address
    ) -> Option<Ballot>;
    fn get_voter(
        &self,
        voting_id: VotingId,
        voting_type: VotingType,
        at: u32
    ) -> Option<Address>;
    fn get_cspr_balance(&self) -> U512;
    fn job_offers_count(&self) -> u32;
    fn jobs_count(&self) -> u32;
    fn bids_count(&self) -> u32;
    fn cancel_voter(&mut self, voter: Address, voting_id: VotingId);
    fn cancel_job_offer(&mut self, job_offer_id: JobOfferId);
    fn cancel_job(&mut self, job_id: JobId);
    fn slash_all_active_job_offers(&mut self, bidder: Address);
    fn slash_bid(&mut self, bid_id: BidId);
    fn voting_exists(&self, voting_id: VotingId, voting_type: VotingType) -> bool;
    fn slash_voter(&mut self, voter: Address, voting_id: VotingId);
    fn change_ownership(&mut self, owner: Address);
    fn add_to_whitelist(&mut self, address: Address);
    fn remove_from_whitelist(&mut self, address: Address);
    fn is_whitelisted(&self, address: Address) -> bool;
    fn get_owner(&self) -> Option<Address>;

}

Defines the Bid Escrow Contract contract’s public interface.

Required Methods

fn init(
&mut self,
variable_repository: Address,
reputation_token: Address,
kyc_token: Address,
va_token: Address
)

Constructor function.

Note

Initializes contract elements:

Sets up ContractRefsWithKycStorage by writing addresses of Variable Repository, Reputation Token, VA Token, KYC Token.

Sets caller as the owner of the contract.

Adds caller to the whitelist.

Events

Emits:

OwnerChanged,

AddedToWhitelist,

fn post_job_offer(
&mut self,
expected_timeframe: BlockTime,
budget: U512,
purse: URef
)

Job Poster post a new Job Offer.

Parameters:

expected_timeframe - Expected timeframe for completing a Job

budget - Maximum budget for a Job Alongside Job Offer, Job Poster also sends DOS Fee in CSPR

Events

JobOfferCreated

fn submit_bid(
&mut self,
job_offer_id: JobOfferId,
time: BlockTime,
payment: U512,
reputation_stake: U512,
onboard: bool,
purse: Option<URef>
)

Worker submits a Bid for a Job.

Parameters:

time - proposed timeframe for completing a Job

payment - proposed payment for a Job

reputation_stake - reputation stake for a Job if Worker is an Internal Worker

onboard - if Worker is an External Worker, then Worker can request to be onboarded after completing a Job

purse: purse containing stake from External Worker

Events

BidSubmitted

fn cancel_bid(&mut self, bid_id: BidId)

Worker cancels a Bid for a Job.

Bid can be cancelled only after VABidAcceptanceTimeout time has passed after submitting a Bid.

Parameters:

bid_id - Bid Id

fn pick_bid(&mut self, job_offer_id: u32, bid_id: u32, purse: URef)

Job poster picks a bid. This creates a new Job object and saves it in a storage. If worker is not onboarded, the job is accepted automatically. Otherwise, worker needs to accept job.

Events

JobCreated

JobAccepted

Errors

CannotPostJobForSelf when trying to create job for self.

JobPosterNotKycd or Error::WorkerNotKycd When either Job Poster or Worker has not completed the KYC process

fn submit_job_proof(&mut self, job_id: JobId, proof: DocumentHash)

Submits a job proof. This is called by a Worker or any KYC’d user during Grace Period. This starts a new voting over the result.

Events

JobSubmitted

Errors

Throws JobAlreadySubmitted if job was already submitted. Throws NotAuthorizedToSubmitResult if one of the constraints for job submission is not met.

fn submit_job_proof_during_grace_period(
&mut self,
job_id: JobId,
proof: DocumentHash,
reputation_stake: U512,
onboard: bool,
purse: Option<URef>
)

Updates the old Bid and Job, the job is assigned to a new Worker. The rest goes the same as regular proof submission. See submit_job_proof(). The old Worker who didn’t submit the proof in time, is getting slashed.

fn vote(
&mut self,
voting_id: VotingId,
voting_type: VotingType,
choice: Choice,
stake: U512
)

Casts a vote over a job.

Events

BallotCast

Errors

CannotVoteOnOwnJob if the voter is either of Job Poster or Worker

VotingNotStarted if the voting was not yet started for this job

fn finish_voting(&mut self, voting_id: VotingId, voting_type: VotingType)

Finishes voting. Depending on type of voting, different actions are performed. Read more

fn get_job(&self, job_id: JobId) -> Option<Job>

Returns a job with given JobId.

fn get_job_offer(&self, job_offer_id: JobOfferId) -> Option<JobOffer>

Returns a JobOffer with given JobOfferId.

fn get_bid(&self, bid_id: BidId) -> Option<Bid>

Returns a Bid with given BidId.

fn variable_repository_address(&self) -> Address

Returns the address of Variable Repository contract.

fn reputation_token_address(&self) -> Address

Returns the address of Reputation Token contract.

fn get_voting(&self, voting_id: VotingId) -> Option<VotingStateMachine>

Returns Voting for given id.

fn get_ballot(
&self,
voting_id: VotingId,
voting_type: VotingType,
address: Address
) -> Option<Ballot>

Returns the Voter’s Ballot.

fn get_voter(
&self,
voting_id: VotingId,
voting_type: VotingType,
at: u32
) -> Option<Address>

Returns the address of nth voter who voted on Voting with voting_id.

fn get_cspr_balance(&self) -> U512

Returns the CSPR balance of the contract.

fn job_offers_count(&self) -> u32

Returns the total number of job offers.

fn jobs_count(&self) -> u32

Returns the total number of jobs.

fn bids_count(&self) -> u32

Returns the total number of bids.

fn cancel_voter(&mut self, voter: Address, voting_id: VotingId)

Erases the voter from voting with the given id. Read more.

fn cancel_job_offer(&mut self, job_offer_id: JobOfferId)

Invalidates the Job Offer, returns DOS Fee to the Job Poster, returns funds to Bidders. Read more.

fn cancel_job(&mut self, job_id: JobId)

Terminates the Voting process and slashes the Worker. Read more.

fn slash_all_active_job_offers(&mut self, bidder: Address)

Invalidates all the active JobOffers, returns DOS Fees to the Job Posters, returns funds to Bidders.

Only a whitelisted account is permitted to call this method. Interacts with Reputation Token Contract.

Errors

Error::BidNotFound

Error::JobOfferNotFound

Error::CannotCancelBidOnCompletedJobOffer

fn slash_bid(&mut self, bid_id: BidId)

Updates the Bid status and returns locked reputation to the Bidder.

Only a whitelisted account is permitted to call this method. Interacts with Reputation Token Contract.

Errors

Error::BidNotFound

Error::JobOfferNotFound

Error::CannotCancelBidOnCompletedJobOffer

fn voting_exists(&self, voting_id: VotingId, voting_type: VotingType) -> bool

Checks if voting of a given type and id exists.

fn slash_voter(&mut self, voter: Address, voting_id: VotingId)

Erases the voter from the given voting. Read more.

fn change_ownership(&mut self, owner: Address)

Changes the ownership of the contract. Transfers ownership to the owner. Only the current owner is permitted to call this method. Read more

fn add_to_whitelist(&mut self, address: Address)

Adds a new address to the whitelist. Read more

fn remove_from_whitelist(&mut self, address: Address)

Remove address from the whitelist. Read more

fn is_whitelisted(&self, address: Address) -> bool

Checks whether the given address is added to the whitelist. Read more.

fn get_owner(&self) -> Option<Address>

Returns the address of the current owner. Read more.