Module casper_dao_contracts::ids

Contains Dao Ids Contract definition and related abstractions.

There is one continuous indexation of votes in the system. Each new voting gets a unique across-the-system id generated by the contract.

Structs

DaoIdsContract

Dao Ids contract manages voting ids in the system. Only a whitelisted account is eligible to generate ids.

DaoIdsContractCaller

Provides a reference to a deployed Dao Ids Contract .

DaoIdsContractTest

A wrapper around DaoIdsContract to simplify testing. Implements DaoIdsContractInterface and TestContract.

Traits

DaoIdsContractInterface

Defines the Dao Ids Contract contract’s public interface.