Trait casper_dao_contracts::ids::DaoIdsContractInterface

pub trait DaoIdsContractInterface {
    fn init(&mut self);
    fn next_voting_id(&mut self) -> VotingId;
    fn change_ownership(&mut self, owner: Address);
    fn add_to_whitelist(&mut self, address: Address);
    fn remove_from_whitelist(&mut self, address: Address);
    fn is_whitelisted(&self, address: Address) -> bool;
    fn get_owner(&self) -> Option<Address>;
}

Defines the Dao Ids Contract contract’s public interface.

Required Methods

fn init(&mut self)

Initializes a contract. Sets the deployer as the owner.

see AccessControl

fn next_voting_id(&mut self) -> VotingId

Returns the next voting id in the system.

Errors

Throws NotWhitelisted if the caller is not whitelisted.

fn change_ownership(&mut self, owner: Address)

Changes the ownership of the contract. Transfers the ownership to the owner. Only the current owner is permitted to call this method.

Read more

fn add_to_whitelist(&mut self, address: Address)

Adds a new address to the whitelist.

Read more

fn remove_from_whitelist(&mut self, address: Address)

Remove address from the whitelist.

Read more

fn is_whitelisted(&self, address: Address) -> bool

Checks whether the given address is added to the whitelist.

Read more.

fn get_owner(&self) -> Option<Address>

Returns the address of the current owner.

Read more.