Module casper_dao_contracts::kyc_nft

Contains KYC NFT Contract definition and related abstractions.

Definitions

KYC - Know Your Customer, is a process that validates that the user can be the user of the system.

Purpose

Ownership of a KYC token indicates the address has been successfully verified and is eligible to participate in the system. Minting token is usually done as a result of KYC Voting.

Each Address can own only one KYC token.

Structs

KycNftContract

Kyc Owned Nft contract acts like an erc-721 token and derives most of erc-721 standards from ERC721Token module.

KycNftContractCaller

Provides a reference to a deployed Kyc Nft Contract .

KycNftContractTest

A wrapper around KycNftContract to simplify testing. Implements KycNftContractInterface and TestContract.

Traits

KycNftContractInterface

Defines the Kyc Nft Contract contract’s public interface.