Module casper_dao_contracts::kyc_voter

Contains KYC Voter Contract definition and related abstractions.

Definitions

KYC - Know Your Customer, is a process that validates that the user can be the user of the system.

KYC process

A type of Governance Voting which the VAs validate a user KYC submission. If VAs vote in favor, the address is considered verified and the KYC Token contract is called in order to mint a token which formally ends the KYC process.

Voting

The Voting process is managed by VotingEngine. There can be only one active voting for a given subject address. An address that is already KYC’d cannot be the subject of voting.

Structs

KycVoterContract

KycVoterContract

KycVoterContractCaller

Provides a reference to a deployed Kyc Voter Contract .

KycVoterContractTest

A wrapper around KycVoterContract to simplify testing. Implements KycVoterContractInterface and TestContract.

KycVotingCreated

Informs kyc voting has been created.

Traits

KycVoterContractInterface

Defines the Kyc Voter Contract contract’s public interface.