Trait casper_dao_contracts::kyc_voter::KycVoterContractInterface

pub trait KycVoterContractInterface {

    fn init(
        &mut self,
        variable_repository: Address,
        reputation_token: Address,
        va_token: Address,
        kyc_token: Address
    );
    fn create_voting(
        &mut self,
        subject_address: Address,
        document_hash: DocumentHash,
        stake: U512
    );
    fn vote(
        &mut self,
        voting_id: VotingId,
        voting_type: VotingType,
        choice: Choice,
        stake: U512
    );
    fn finish_voting(&mut self, voting_id: VotingId, voting_type: VotingType);
    fn variable_repository_address(&self) -> Address;
    fn reputation_token_address(&self) -> Address;
    fn voting_exists(&self, voting_id: VotingId, voting_type: VotingType) -> bool;
    fn slash_voter(&mut self, voter: Address, voting_id: VotingId);
    fn get_voting(&self, voting_id: VotingId) -> Option<VotingStateMachine>;
    fn get_ballot(
        &self,
        voting_id: VotingId,
        voting_type: VotingType,
        address: Address
    ) -> Option<Ballot>;
    fn get_voter(
        &self,
        voting_id: VotingId,
        voting_type: VotingType,
        at: u32
    ) -> Option<Address>;
    fn kyc_token_address(&self) -> Address;
    fn change_ownership(&mut self, owner: Address);
    fn add_to_whitelist(&mut self, address: Address);
    fn remove_from_whitelist(&mut self, address: Address);
    fn is_whitelisted(&self, address: Address) -> bool;
    fn get_owner(&self) -> Option<Address>;

}

Defines the Kyc Voter Contract contract’s public interface.

Required Methods

fn init(
&mut self,
variable_repository: Address,
reputation_token: Address,
va_token: Address,
kyc_token: Address
)

Constructor function.

Note

Initializes contract elements:

Sets up ContractRefsWithKycStorage by writing addresses of Variable Repository, Reputation Token, VA Token, KYC Token.

Sets caller as the owner of the contract.

Adds caller to the whitelist.

Events

OwnerChanged,

AddedToWhitelist,

fn create_voting(
&mut self,
subject_address: Address,
document_hash: DocumentHash,
stake: U512
)

Creates a new KYC Voting. Once the voting passes a KYC Token is minted to the subject_address.

Prerequisites

no voting on the given subject_address is in progress,

subject_address does not own a kyc token.

Arguments

subject_address - address of a user to be verified,

document_hash - a hash of a document that verifies the user. The hash is used as an id of a freshly minted kyc token,

subject_address - an Address to be on/offboarded.

Events

KycVotingCreated

fn vote(
&mut self,
voting_id: VotingId,
voting_type: VotingType,
choice: Choice,
stake: U512
)

Casts a vote. Read more

fn finish_voting(&mut self, voting_id: VotingId, voting_type: VotingType)

Finishes voting. Depending on type of voting, different actions are performed. Read more

fn variable_repository_address(&self) -> Address

Returns the address of Variable Repository contract.

fn reputation_token_address(&self) -> Address

Returns the address of Reputation Token contract.

fn voting_exists(&self, voting_id: VotingId, voting_type: VotingType) -> bool

Checks if voting of a given type and id exists.

fn slash_voter(&mut self, voter: Address, voting_id: VotingId)

Erases the voter from voting with the given id. Read more.

fn get_voting(&self, voting_id: VotingId) -> Option<VotingStateMachine>

Returns Voting for given id.

fn get_ballot(
&self,
voting_id: VotingId,
voting_type: VotingType,
address: Address
) -> Option<Ballot>

Returns the Voter’s Ballot.

fn get_voter(
&self,
voting_id: VotingId,
voting_type: VotingType,
at: u32
) -> Option<Address>

Returns the address of nth voter who voted on Voting with voting_id.

fn kyc_token_address(&self) -> Address

Returns the address of KYC Token contract.

fn change_ownership(&mut self, owner: Address)

Changes the ownership of the contract. Transfers ownership to the owner. Only the current owner is permitted to call this method. Read more

fn add_to_whitelist(&mut self, address: Address)

Adds a new address to the whitelist. Read more

fn remove_from_whitelist(&mut self, address: Address)

Remove address from the whitelist. Read more

fn is_whitelisted(&self, address: Address) -> bool

Checks whether the given address is added to the whitelist. Read more.

fn get_owner(&self) -> Option<Address>

Returns the address of the current owner. Read more.