Trait casper_dao_contracts::onboarding_request::OnboardingRequestContractInterface

pub trait OnboardingRequestContractInterface {

    fn init(
        &mut self,
        variable_repository: Address,
        reputation_token: Address,
        kyc_token: Address,
        va_token: Address
    );
    fn create_voting(&mut self, reason: DocumentHash, purse: URef);
    fn vote(
        &mut self,
        voting_id: VotingId,
        voting_type: VotingType,
        choice: Choice,
        stake: U512
    );
    fn finish_voting(&mut self, voting_id: VotingId, voting_type: VotingType);
    fn variable_repository_address(&self) -> Address;
    fn reputation_token_address(&self) -> Address;
    fn get_voting(&self, voting_id: VotingId) -> Option<VotingStateMachine>;
    fn get_ballot(
        &self,
        voting_id: VotingId,
        voting_type: VotingType,
        address: Address
    ) -> Option<Ballot>;
    fn get_voter(
        &self,
        voting_id: VotingId,
        voting_type: VotingType,
        at: u32
    ) -> Option<Address>;
    fn voting_exists(&self, voting_id: VotingId, voting_type: VotingType) -> bool;
    fn slash_voter(&mut self, voter: Address, voting_id: VotingId);
    fn get_cspr_balance(&self) -> U512;
    fn change_ownership(&mut self, owner: Address);
    fn add_to_whitelist(&mut self, address: Address);
    fn remove_from_whitelist(&mut self, address: Address);
    fn is_whitelisted(&self, address: Address) -> bool;
    fn get_owner(&self) -> Option<Address>;

}

Defines the Onboarding Request Contract contract’s public interface.

Required Methods

fn init(
&mut self,
variable_repository: Address,
reputation_token: Address,
kyc_token: Address,
va_token: Address
)

Constructor function.

Note

Initializes contract elements:

Sets up ContractRefsWithKycStorage by writing addresses of Variable Repository, Reputation Token, VA Token, KYC Token.

Sets caller as the owner of the contract.

Adds caller to the whitelist.

Events

OwnerChanged,

AddedToWhitelist,

fn create_voting(&mut self, reason: DocumentHash, purse: URef)

Submits an onboarding request. If the request is valid voting starts.

Events

OnboardingVotingCreated

fn vote(
&mut self,
voting_id: VotingId,
voting_type: VotingType,
choice: Choice,
stake: U512
)

Casts a vote. Read more

fn finish_voting(&mut self, voting_id: VotingId, voting_type: VotingType)

Finishes voting stage. Depending on stage, the voting can be converted to a formal one, end with a refund or convert the requestor to a VA.

fn variable_repository_address(&self) -> Address

Returns the address of Variable Repository contract.

fn reputation_token_address(&self) -> Address

Returns the address of Reputation Token contract.

fn get_voting(&self, voting_id: VotingId) -> Option<VotingStateMachine>

Returns Voting for given id.

fn get_ballot(
&self,
voting_id: VotingId,
voting_type: VotingType,
address: Address
) -> Option<Ballot>

Returns the Voter’s Ballot.

fn get_voter(
&self,
voting_id: VotingId,
voting_type: VotingType,
at: u32
) -> Option<Address>

Gets the address of nth voter who voted on Voting with voting_id.

fn voting_exists(&self, voting_id: VotingId, voting_type: VotingType) -> bool

Checks if voting of a given type and id exists.

fn slash_voter(&mut self, voter: Address, voting_id: VotingId)

Erases the voter from voting with the given id. Read more.

fn get_cspr_balance(&self) -> U512

Gets the CSPR balance of the contract.

fn change_ownership(&mut self, owner: Address)

Changes the ownership of the contract. Transfers ownership to the owner. Only the current owner is permitted to call this method. Read more

fn add_to_whitelist(&mut self, address: Address)

Adds a new address to the whitelist. Read more

fn remove_from_whitelist(&mut self, address: Address)

Remove address from the whitelist. Read more

fn is_whitelisted(&self, address: Address) -> bool

Checks whether the given address is added to the whitelist. Read more.

fn get_owner(&self) -> Option<Address>

Returns the address of the current owner. Read more.