Module casper_dao_contracts::repo_voter

Contains Repo Voter Contract definition and related abstractions.

General

A type of Governance Voting that updates values stored in the Variable Repository Contract.

Voting

The Voting process is managed by VotingEngine.

Structs

RepoVoterContract

RepoVoterContract

RepoVoterContractCaller

Provides a reference to a deployed Repo Voter Contract .

RepoVoterContractTest

A wrapper around RepoVoterContract to simplify testing. Implements RepoVoterContractInterface and TestContract.

RepoVotingCreated

Informs repo voting has been created.

Traits

RepoVoterContractInterface

Defines the Repo Voter Contract contract’s public interface.