Struct casper_dao_contracts::reputation::ReputationContractCaller

pub struct ReputationContractCaller { /* private fields */ }

Provides a reference to a deployed Reputation Contract .

Implementations

impl ReputationContractCaller

pub fn at(address: Address) -> Self

Creates a new caller instance from the given address.