Trait casper_dao_contracts::reputation::ReputationContractInterface

pub trait ReputationContractInterface {

    fn init(&mut self);
    fn mint(&mut self, recipient: Address, amount: U512);
    fn mint_passive(&mut self, recipient: Address, amount: U512);
    fn burn(&mut self, owner: Address, amount: U512);
    fn burn_passive(&mut self, owner: Address, amount: U512);
    fn change_ownership(&mut self, owner: Address);
    fn add_to_whitelist(&mut self, address: Address);
    fn remove_from_whitelist(&mut self, address: Address);
    fn get_owner(&self) -> Option<Address>;
    fn total_supply(&self) -> U512;
    fn balance_of(&self, address: Address) -> U512;
    fn passive_balance_of(&self, address: Address) -> U512;
    fn is_whitelisted(&self, address: Address) -> bool;
    fn stake_voting(&mut self, voting_id: VotingId, ballot: ShortenedBallot);
    fn unstake_voting(&mut self, voting_id: VotingId, ballot: ShortenedBallot);
    fn bulk_unstake_voting(
        &mut self,
        voting_id: VotingId,
        ballots: Vec<ShortenedBallot>
    );
    fn stake_bid(&mut self, bid: ShortenedBid);
    fn unstake_bid(&mut self, bid: ShortenedBid);
    fn bulk_unstake_bid(&mut self, bids: Vec<ShortenedBid>);
    fn get_stake(&self, address: Address) -> U512;
    fn all_balances(&self) -> AggregatedBalance;
    fn partial_balances(&self, addresses: Vec<Address>) -> AggregatedBalance;
    fn bulk_mint_burn(
        &mut self,
        mints: BTreeMap<Address, U512>,
        burns: BTreeMap<Address, U512>
    );
    fn burn_all(&mut self, owner: Address);
    fn stakes_info(&self, address: Address) -> AggregatedStake;

}

Defines the Reputation Contract contract’s public interface.

Required Methods

fn init(&mut self)

Constructor method.

It initializes contract elements:

Events dictionary.

Named keys of AccessControl.

Set caller as the owner of the contract.

Add caller to the whitelist.

Events

OwnerChanged,

AddedToWhitelist.

fn mint(&mut self, recipient: Address, amount: U512)

Mints new tokens. Adds amount of new tokens to the balance of the recipient and increments the total supply. Only whitelisted addresses are permitted to call this method.

Errors

NotWhitelisted if caller is not whitelisted.

Events

Mint.

fn mint_passive(&mut self, recipient: Address, amount: U512)

Increases the balance of the passive reputation of the given address.

Errors

NotWhitelisted if caller is not whitelisted.

fn burn(&mut self, owner: Address, amount: U512)

Burns existing tokens. Removes amount of existing tokens from the balance of the owner and decrements the total supply. Only whitelisted addresses are permitted to call this method.

Errors

NotWhitelisted if caller is not whitelisted.

Events

Burn event.

fn burn_passive(&mut self, owner: Address, amount: U512)

Decreases the balance of the passive reputation of the given address.

Errors

NotWhitelisted if caller is not whitelisted.

InsufficientBalance if the passed amount exceeds the balance of the passive reputation of the given address.

fn change_ownership(&mut self, owner: Address)

Changes ownership of the contract. Transfer the ownership to the owner. Only the current owner is permitted to call this method.

See AccessControl

fn add_to_whitelist(&mut self, address: Address)

Adds a new address to the whitelist.

See AccessControl

fn remove_from_whitelist(&mut self, address: Address)

Removes address from the whitelist.

See AccessControl

fn get_owner(&self) -> Option<Address>

Returns the address of the current owner.

fn total_supply(&self) -> U512

Returns the total token supply.

fn balance_of(&self, address: Address) -> U512

Returns the current token balance of the given address.

fn passive_balance_of(&self, address: Address) -> U512

Returns the current passive balance of the given address.

fn is_whitelisted(&self, address: Address) -> bool

Checks whether the given address is added to the whitelist.

fn stake_voting(&mut self, voting_id: VotingId, ballot: ShortenedBallot)

Stakes the reputation used as voting power.

fn unstake_voting(&mut self, voting_id: VotingId, ballot: ShortenedBallot)

fn bulk_unstake_voting(
&mut self,
voting_id: VotingId,
ballots: Vec<ShortenedBallot>
)

Unstakes the reputation used as voting power.

fn stake_bid(&mut self, bid: ShortenedBid)

Stakes the reputation used as bid value.

fn unstake_bid(&mut self, bid: ShortenedBid)

Unstakes the reputation used as bid value.

fn bulk_unstake_bid(&mut self, bids: Vec<ShortenedBid>)

Unstakes the reputation used as bid value.

fn get_stake(&self, address: Address) -> U512

Returns the total stake of the given account.

fn all_balances(&self) -> AggregatedBalance

Gets balances of all the token holders.

fn partial_balances(&self, addresses: Vec<Address>) -> AggregatedBalance

Gets balances of the given account addresses.

fn bulk_mint_burn(
&mut self,
mints: BTreeMap<Address, U512>,
burns: BTreeMap<Address, U512>
)

Redistributes the reputation based on the voting summary

fn burn_all(&mut self, owner: Address)

Burns all the tokens of the owner.

fn stakes_info(&self, address: Address) -> AggregatedStake

Returns all the data about the given user stakes.