Module casper_dao_contracts::reputation_voter

Contains Reputation Voter Contract definition and related abstractions.

General

A type of Governance Voting used to operate on the Reputation Token Contract.

Two types of voting can be created:

to mint tokens for a user,

to burn users’ tokens.

Voting

The Voting process is managed by VotingEngine.

Structs

ReputationVoterContract

ReputationVoterContract

ReputationVoterContractCaller

Provides a reference to a deployed Reputation Voter Contract .

ReputationVoterContractTest

A wrapper around ReputationVoterContract to simplify testing. Implements ReputationVoterContractInterface and TestContract.

ReputationVoting

Struct storing all information about reputation voting

ReputationVotingCreated

Informs reputation voting has been created.

Enums

Action

Action to perform against reputation

Traits

ReputationVoterContractInterface

Defines the Reputation Voter Contract contract’s public interface.