Trait casper_dao_contracts::va_nft::VaNftContractInterface

pub trait VaNftContractInterface {

    fn init(&mut self, name: String, symbol: String, base_uri: TokenUri);
    fn change_ownership(&mut self, owner: Address);
    fn add_to_whitelist(&mut self, address: Address);
    fn remove_from_whitelist(&mut self, address: Address);
    fn is_whitelisted(&self, address: Address) -> bool;
    fn get_owner(&self) -> Option<Address>;
    fn name(&self) -> String;
    fn symbol(&self) -> String;
    fn owner_of(&self, token_id: TokenId) -> Option<Address>;
    fn token_id(&self, address: Address) -> Option<TokenId>;
    fn balance_of(&self, owner: Address) -> U512;
    fn total_supply(&self) -> U512;
    fn token_uri(&self, token_id: TokenId) -> TokenUri;
    fn base_uri(&self) -> TokenUri;
    fn mint(&mut self, to: Address);
    fn burn(&mut self, owner: Address);

}

Defines the Va Nft Contract contract’s public interface.

Required Methods

fn init(&mut self, name: String, symbol: String, base_uri: TokenUri)

Contract constructor.

Initializes modules. Sets the deployer as the owner.

See MetadataERC721, AccessControl

fn change_ownership(&mut self, owner: Address)

Changes the ownership of the contract. Transfers ownership to the owner. Only the current owner is permitted to call this method. Read more

fn add_to_whitelist(&mut self, address: Address)

Adds a new address to the whitelist. Read more

fn remove_from_whitelist(&mut self, address: Address)

Remove address from the whitelist. Read more

fn is_whitelisted(&self, address: Address) -> bool

Checks whether the given address is added to the whitelist. Read more.

fn get_owner(&self) -> Option<Address>

Returns the address of the current owner. Read more.

fn name(&self) -> String

Returns a descriptive name for a collection of tokens in this contract.

fn symbol(&self) -> String

Gets an abbreviated name for tokens in this contract.

fn owner_of(&self, token_id: TokenId) -> Option<Address>

Returns the address of the owner of the token.

If the given token_id does not exist the None value is returned.

fn token_id(&self, address: Address) -> Option<TokenId>

Returns a token id for the given the address.

If the owner does not own any token the None value is returned.

fn balance_of(&self, owner: Address) -> U512

Returns the number of tokens owned by owner.

fn total_supply(&self) -> U512

Returns the total number of tokens.

fn token_uri(&self, token_id: TokenId) -> TokenUri

Returns a distinct Uniform Resource Identifier (URI) for a given asset.

fn base_uri(&self) -> TokenUri

Returns a URI prefix that is used by all the assets.

fn mint(&mut self, to: Address)

Creates a new token with the next id and transfers it to a new owner. Increments the total supply and the balance of the to address.

Note

Only whitelisted addresses are permitted to call this method.

Each user is entitled to own only one token.

Errors

UserAlreadyOwnsToken if the to address already owns a token.

Events

Transfer when minted successfully.

fn burn(&mut self, owner: Address)

Burns a token with a given id. Decrements the balance of the token owner and decrements the total supply.

Errors

NotWhitelisted if caller is not whitelisted.

Events

Transfer when burnt successfully.