Module casper_dao_contracts::variable_repository

Contains Variable Repository Contract definition and related abstractions.

Variable Repository Contract stores governance variables. Values can be altered as a result of Repo Voting.

Available keys

Parameter name	Initial value	Stored value	Type	Description

PostJobDOSFee	10	10000	float	A DOS fee that the JobPoster needs to attach to the Post Job query. The value is the minimum amount of Fiat currency to be attached as CSPR using FiatConversionRate

InternalAuctionTime	7 days	604800	seconds	The time of the Internal Auction

PublicAuctionTime	10 days	864000	seconds	The time of the External Auction

DefaultPolicingRate	0.3	300	float	This rate defines how many Reputation tokens are given to the VA’s for their community audit/vote on a work product.. In case of value of 0.3, 30% of the payment is redistributed between VA’s and 70% is given to the Worker.

ReputationConversionRate	0.1	100	float	This parameter tells how much Reputation is minted for each unit of currency paid for Work. For value equal to 0.1, 1 Reputation is minted for each 10 CSPR.

FiatConversionRateAddress			address	An address of a contract that will return the conversion rate between Fiat and CSPR

ForumKYCRequired	true	true	bool	Defines if KYC is required to post on Forum

BidEscrowInformalQuorumRatio	0.5	500	float	How many holders of the Reputation tokens (VA’s) are needed for an informal voting quorum

BidEscrowFormalQuorumRatio	0.5	500	float	How many holders of the Reputation tokens (VA’s) are needed for a bid escrow formal vote quorum. For example, if 100 accounts hold tokens, the quorum would be 51 votes.

InformalQuorumRatio	0.5	500	float	How many holders of the Reputation tokens (VA’s) are needed for a regular informal voting quorum.

FormalQuorumRatio	0.5	500	float	How many holders of the Reputation tokens (VA’s) are needed for a regular formal voting quorum.

BidEscrowInformalVotingTime	5 days	432000	seconds	Time for the informal part of the Bid Escrow voting

BidEscrowFormalVotingTime	5 days	432000	seconds	Time for the formal part of the Bid Escrow voting

InformalVotingTime	5 days	432000	seconds	Time for the informal part of other voting

FormalVotingTime	5 days	432000	seconds	Time for the formal part of other voting

InformalStakeReputation	true	true	bool	Tells if the Informal Voting should stake the reputation or only simulate it.

TimeBetweenInformalAndFormalVoting	1 day	86400	seconds	Time between Informal and Formal Voting

VABidAcceptanceTimeout	2 days	172800	seconds	How much time the bid wait for the acceptance. After this time, the bid can be cancelled

VACanBidOnPublicAuction	false	false	bool	Whether or not VA’s can take part in the Public Auction part of the Bidding process.

DistributePaymentToNonVoters	true	false	bool	Determines if the Payment for the Job should be distributed between all VA’s or only to those who voted

DefaultReputationSlash	0.1	100	float	How much reputation of an Internal Worker is slashed after not completing a Job

VotingClearnessDelta	8	8000	int	If the difference between 50/50 and result of the Informal Voting is bigger than the value, the time between votings should be doubled.

VotingStartAfterJobWorkerSubmisson	3 days	259200	seconds	Time between the worker job submission and the internal voting start.

BidEscrowPaymentRatio	0.1	100	float	How much CSPR is sent to GovernanceWallet after the Job is finished

BidEscrowWalletAddress			address	An address of a multisig wallet of the DAO.

Structs

VariableRepositoryContract

Variable Repository Contract implementation. See VariableRepositoryContractInterface.

VariableRepositoryContractCaller

Provides a reference to a deployed Variable Repository Contract .

VariableRepositoryContractTest

A wrapper around VariableRepositoryContract to simplify testing. Implements VariableRepositoryContractInterface and TestContract.

Traits

VariableRepositoryContractInterface

Defines the Variable Repository Contract contract’s public interface.