Trait casper_dao_contracts::variable_repository::VariableRepositoryContractInterface

pub trait VariableRepositoryContractInterface {
    fn init(&mut self);
    fn change_ownership(&mut self, owner: Address);
    fn add_to_whitelist(&mut self, address: Address);
    fn remove_from_whitelist(&mut self, address: Address);
    fn is_whitelisted(&self, address: Address) -> bool;
    fn get_owner(&self) -> Option<Address>;
    fn update_at(
        &mut self,
        key: String,
        value: Bytes,
        activation_time: Option<u64>
    );
    fn get(&self, key: String) -> Option<Bytes>;
    fn get_full_value(&self, key: String) -> Option<Record>;
    fn get_key_at(&self, index: u32) -> Option<String>;
    fn keys_count(&self) -> u32;
    fn all_variables(&self) -> BTreeMap<String, Bytes>;
}

Defines the Variable Repository Contract contract’s public interface.

Required Methods

fn init(&mut self)

Constructor function.

Note

Initializes contract elements:

Sets the default configuration of the Repository

Sets caller as the owner of the contract.

Adds caller to the whitelist.

Events

OwnerChanged,

AddedToWhitelist,

multiple ValueUpdated events, one per value of the default repository configuration.

fn change_ownership(&mut self, owner: Address)

Changes the ownership of the contract. Transfers ownership to the owner. Only the current owner is permitted to call this method. Read more

fn add_to_whitelist(&mut self, address: Address)

Adds a new address to the whitelist. Read more

fn remove_from_whitelist(&mut self, address: Address)

Remove address from the whitelist. Read more

fn is_whitelisted(&self, address: Address) -> bool

Checks whether the given address is added to the whitelist. Read more.

fn get_owner(&self) -> Option<Address>

Returns the address of the current owner. Read more.

fn update_at(&mut self, key: String, value: Bytes, activation_time: Option<u64>)

Inserts or updates the value under the given key.

Note

The activation time is represented as a unix timestamp.

If the activation time is None the value is updated immediately.

If some future time in the future is passed as an argument, the Self::get function returns the previously set value.

Events

ValueUpdated.

Errors

NotWhitelisted if the caller is not a whitelisted user.

ActivationTimeInPast if the activation time has passed already.

ValueNotAvailable on the future value update if the current value has not been set.

fn get(&self, key: String) -> Option<Bytes>

Returns the value stored under the given key.

If the key does not exist, the None value is returned.

fn get_full_value(&self, key: String) -> Option<Record>

Returns the full (current and future) value stored under the given key. See Record.

If the key does not exist, the None value is returned.

fn get_key_at(&self, index: u32) -> Option<String>

Returns the value stored under the given index.

Every freshly added key has the previous key index increased by 1. The index range is 0 to #keys-1.

If the given index exceeds #keys-1 the None value is returned.

fn keys_count(&self) -> u32

Returns the number of existing keys in the Repository.

fn all_variables(&self) -> BTreeMap<String, Bytes>

Reads all the stored variables and returns a map key to value.