Module casper_dao_contracts::voting

Voting utilities.

Choice

Voting is a binary choice. Each voter may vote:

In favor - yes vote

Against - no vote.

Calculating Quorum

During both types of Voting a specific amount of votes is required to be cast - the Quorum. The proportional amount of votes is defined in the configuration:

Informal Quorum Ratio,

Formal Quorum Ratio.

To calculate the amount of votes required to achieve quorum we use the following formula:

quorum = total amount of VAs' * QuorumRatio

For example, with 11 VA’s and QuorumRatio set to 0.5, the quorum is 5.5, which means that there need to be at least 6 votes cast for the Voting.

Informal Voting

The Informal Voting is the first phase of the Voting process. Its parameters are configurable:

Informal Voting Time - how long the voting lasts,

Informal Quorum Ratio - how many VA votes are needed,

Informal Stake Reputation - if the Reputation used for Informal Voting should be staked or not.

The first vote in the Informal Voting may be casted automatically at creation time as a yes depending on configuration.

Voting passed

When Informal Voting passes, following things happen:

The time between Votes starts to count,

Creator’s stake is used as a yes vote for the first vote in the Formal Voting,

VAs' stakes are returned to them.

Voting failed

When Informal Voting fails, following things happen:

The time between Votes starts to count,

Creator’s stake is used as a yes vote for the first vote in the Formal Voting,

VAs' stakes are returned to them.

Quorum not reached

When the Quorum is not reached during the Informal Voting, following things happen:

The process ends here,

VAs' stakes are returned to them,

Time between Votes

After passing or failing Informal Voting, there is a certain amount of time before the Formal Voting starts. However, depending on the vote difference in the Informal Voting this time can be doubled.

Formal Voting

The Formal Voting is the final step in the Voting process. Its parameters are configured using governance variables:

Voting Time - how long the voting lasts,

Quorum Ratio - how many VA votes are needed,

Different actions are performed by the contract depending on the result.

Voting passed

Besides yielding a positive result, the Voting passed means that the Reputation staked by the losing side is redistributed between the winning side proportionally to the amount of reputation staked in the voting.

Voting failed

Besides yielding a negative result, the Voting passed means that the Reputation staked by the losing side is redistributed between the winning side proportionally to the amount of reputation staked in the voting.

Quorum not reached

When the Quorum is not reached during the Formal Voting, following things happen:

The process ends here.

VAs' stakes are returned to them.

Modules

events

Set of voting events.

refs

Utility modules providing references to common contracts that are used by most of the voting contracts.

submodules

Voting utility submodules.

voting_state_machine

Voting struct with logic for governance voting

Structs

Ballot

Represents user’s vote.

ShortenedBallot

Short version of Ballot struct.

VotingEngine

Governance voting is a struct that contracts can use to implement voting.

Enums

Choice

Choice enum, can be converted to bool using is_in_favor()

Functions

redistribute_cspr_to_all_vas

Transfers CSPRs to all VAs’. Each VA gets the amount of CSPR proportionally to their reputation.

redistribute_to_governance

Transfers some part of a given amount to Bid Escrow Wallet and returns the remaining amount.

Type Definitions

VotingId

A unique voting identifier.