Module casper_dao_contracts::voting::refs

Utility modules providing references to common contracts that are used by most of the voting contracts.

Structs

ContractRefsStorage

A module that stores addresses to common contracts that are used by most of the voting contracts.

ContractRefsWithKycStorage

A decorated ContractRefsStorage module that additionally stores addresses an address of KYC Token.

Traits

ContractRefs

Provides references to common contracts that are used by most of the voting contracts.