Trait casper_dao_contracts::voting::refs::ContractRefs

pub trait ContractRefs {
    fn reputation_token(&self) -> ReputationContractCaller;
    fn variable_repository(&self) -> VariableRepositoryContractCaller;
    fn va_token(&self) -> VaNftContractCaller;
}

Provides references to common contracts that are used by most of the voting contracts.

Required Methods

fn reputation_token(&self) -> ReputationContractCaller

Returns a reference to Reputation Token connected to the contract

fn variable_repository(&self) -> VariableRepositoryContractCaller

Returns a reference to Variable Repository connected to the contract

fn va_token(&self) -> VaNftContractCaller

Returns a reference to VA Token connected to the contract